entity

+++ recipify_back/recipify_back/src/main/java/org/example/recipify_back/entity/Recipe.java
@@ -18,7 +18,7 @@ public class Recipe {
     private int id;
     private String title;
     private String insctruction;
-    private long calorie;
+    private int calorie;
     private int serving;
     private boolean is_approved;
     private boolean is_private;
